+++ datanode/api/trading_data_v2.go
@@ -1401,6 +1401,39 @@ func (t *tradingDataServiceV2) ListRewardSummaries(ctx context.Context, in *v2.L
 	return &resp, nil
 }
 
+// Get reward summaries for epoch range.
+func (t *tradingDataServiceV2) ListEpochRewardSummaries(ctx context.Context, in *v2.ListEpochRewardSummariesRequest) (*v2.ListEpochRewardSummariesResponse, error) {
+	defer metrics.StartAPIRequestAndTimeGRPC("ListEpochRewardSummaries")()
+
+	if in == nil {
+		return nil, apiError(codes.InvalidArgument, fmt.Errorf("empty request"))
+	}
+
+	pagination, err := entities.CursorPaginationFromProto(in.Pagination)
+	if err != nil {
+		return nil, apiError(codes.InvalidArgument, err)
+	}
+
+	summaries, pageInfo, err := t.rewardService.GetEpochRewardSummaries(ctx, in.FromEpoch, in.ToEpoch, pagination)
+	if err != nil {
+		return nil, apiError(codes.Internal, err)
+	}
+
+	edges, err := makeEdges[*v2.EpochRewardSummaryEdge](summaries)
+	if err != nil {
+		return nil, apiError(codes.Internal, err)
+	}
+
+	connection := v2.EpochRewardSummaryConnection{
+		Edges:    edges,
+		PageInfo: pageInfo.ToProto(),
+	}
+
+	return &v2.ListEpochRewardSummariesResponse{
+		Summaries: &connection,
+	}, nil
+}
+
 // subscribe to rewards.
 func (t *tradingDataServiceV2) ObserveRewards(req *v2.ObserveRewardsRequest, srv v2.TradingDataService_ObserveRewardsServer) error {
 	ctx, cfunc := context.WithCancel(srv.Context())