+++ protos/data-node/api/v2/trading_data_grpc.pb.go
@@ -199,6 +199,8 @@ type TradingDataServiceClient interface {
 	//
 	// Get reward summaries
 	ListRewardSummaries(ctx context.Context, in *ListRewardSummariesRequest, opts ...grpc.CallOption) (*ListRewardSummariesResponse, error)
+	// List reward summaries by epoch
+	ListEpochRewardSummaries(ctx context.Context, in *ListEpochRewardSummariesRequest, opts ...grpc.CallOption) (*ListEpochRewardSummariesResponse, error)
 	// subscribe to rewards
 	ObserveRewards(ctx context.Context, in *ObserveRewardsRequest, opts ...grpc.CallOption) (TradingDataService_ObserveRewardsClient, error)
 	// Deposit
@@ -976,6 +978,15 @@ func (c *tradingDataServiceClient) ListRewardSummaries(ctx context.Context, in *
 	return out, nil
 }
 
+func (c *tradingDataServiceClient) ListEpochRewardSummaries(ctx context.Context, in *ListEpochRewardSummariesRequest, opts ...grpc.CallOption) (*ListEpochRewardSummariesResponse, error) {
+	out := new(ListEpochRewardSummariesResponse)
+	err := c.cc.Invoke(ctx, "/datanode.api.v2.TradingDataService/ListEpochRewardSummaries", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
 func (c *tradingDataServiceClient) ObserveRewards(ctx context.Context, in *ObserveRewardsRequest, opts ...grpc.CallOption) (TradingDataService_ObserveRewardsClient, error) {
 	stream, err := c.cc.NewStream(ctx, &TradingDataService_ServiceDesc.Streams[10], "/datanode.api.v2.TradingDataService/ObserveRewards", opts...)
 	if err != nil {
@@ -1645,6 +1656,8 @@ type TradingDataServiceServer interface {
 	//
 	// Get reward summaries
 	ListRewardSummaries(context.Context, *ListRewardSummariesRequest) (*ListRewardSummariesResponse, error)
+	// List reward summaries by epoch
+	ListEpochRewardSummaries(context.Context, *ListEpochRewardSummariesRequest) (*ListEpochRewardSummariesResponse, error)
 	// subscribe to rewards
 	ObserveRewards(*ObserveRewardsRequest, TradingDataService_ObserveRewardsServer) error
 	// Deposit
@@ -1924,6 +1937,9 @@ func (UnimplementedTradingDataServiceServer) ListRewards(context.Context, *ListR
 }
 func (UnimplementedTradingDataServiceServer) ListRewardSummaries(context.Context, *ListRewardSummariesRequest) (*ListRewardSummariesResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListRewardSummaries not implemented")
+}
+func (UnimplementedTradingDataServiceServer) ListEpochRewardSummaries(context.Context, *ListEpochRewardSummariesRequest) (*ListEpochRewardSummariesResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method ListEpochRewardSummaries not implemented")
 }
 func (UnimplementedTradingDataServiceServer) ObserveRewards(*ObserveRewardsRequest, TradingDataService_ObserveRewardsServer) error {
 	return status.Errorf(codes.Unimplemented, "method ObserveRewards not implemented")
@@ -2877,6 +2893,24 @@ func _TradingDataService_ListRewardSummaries_Handler(srv interface{}, ctx contex
 	return interceptor(ctx, in, info, handler)
 }
 
+func _TradingDataService_ListEpochRewardSummaries_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(ListEpochRewardSummariesRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(TradingDataServiceServer).ListEpochRewardSummaries(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/datanode.api.v2.TradingDataService/ListEpochRewardSummaries",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(TradingDataServiceServer).ListEpochRewardSummaries(ctx, req.(*ListEpochRewardSummariesRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 func _TradingDataService_ObserveRewards_Handler(srv interface{}, stream grpc.ServerStream) error {
 	m := new(ObserveRewardsRequest)
 	if err := stream.RecvMsg(m); err != nil {
@@ -3744,6 +3778,10 @@ var TradingDataService_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "ListRewardSummaries",
 			Handler:    _TradingDataService_ListRewardSummaries_Handler,
+		},
+		{
+			MethodName: "ListEpochRewardSummaries",
+			Handler:    _TradingDataService_ListEpochRewardSummaries_Handler,
 		},
 		{
 			MethodName: "GetDeposit",